psycho_Test/requirements.txt
asgiref==3.8.1
boto3==1.35.54
botocore==1.35.54
colorama==0.4.6
Django==5.1.2
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
loguru==0.7.2
mysql==0.0.3
mysqlclient==2.2.5
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
win32-setctime==1.1.0
